Validar permissão de acesso - Aula 1, Validar permissão de acesso - Aula 2, Ocultar botão da página que o nível de acesso não tem permissão de acesso

+++ app/adms/Views/AccessLevels/ListAccessLevels.php
@@ -13,8 +13,14 @@ if (!defined('G9C8O7N6N5T4I')) {
             <span class="title-content">Listar Níveis de Acesso</span>
             <div class="top-list-right">
                 <?php
-                echo "<a href='" . URLADM . "add-access-level/index' class='btn-success'>Cadastrar</a> ";
-                echo "<a href='" . URLADM . "sync-pages-levels/index' class='btn-warning'>Sincronizar</a>";
+
+                if ($this->data['button']['add_access_level']) {
+                    echo "<a href='" . URLADM . "add-access-level/index' class='btn-success'>Cadastrar</a> ";
+                }
+
+                if ($this->data['button']['sync_pages_levels']) {
+                    echo "<a href='" . URLADM . "sync-pages-levels/index' class='btn-warning'>Sincronizar</a>";
+                }
                 ?>
             </div>
         </div>
@@ -49,11 +55,21 @@ if (!defined('G9C8O7N6N5T4I')) {
                                 <button onclick="actionDropdown(<?php echo $id; ?>)" class="dropdown-btn-action">Ações</button>
                                 <div id="actionDropdown<?php echo $id; ?>" class="dropdown-action-item">
                                     <?php
-                                    echo "<a href='" . URLADM . "order-access-level/index/$id?pag=" . $this->data['pag'] . "'><i class='fa-solid fa-arrow-up-short-wide'></i> Ordem</a>";
-                                    echo "<a href='" . URLADM . "list-permission/index/?level=$id'><i class='fa-solid fa-house-lock'></i> Permissão</a>";
-                                    echo "<a href='" . URLADM . "view-access-level/index/$id'><i class='fa-solid fa-eye'></i> Visualizar</a>";
-                                    echo "<a href='" . URLADM . "edit-access-level/index/$id'><i class='fa-solid fa-pen-to-square'></i> Editar</a>";
-                                    echo "<a href='" . URLADM . "delete-access-level/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")'><i class='fa-regular fa-trash-can'></i> Apagar</a>";
+                                    if ($this->data['button']['view_access_level']) {
+                                        echo "<a href='" . URLADM . "view-access-level/index/$id'><i class='fa-solid fa-eye'></i> Visualizar</a>";
+                                    }
+                                    if ($this->data['button']['edit_access_level']) {
+                                        echo "<a href='" . URLADM . "edit-access-level/index/$id'><i class='fa-solid fa-pen-to-square'></i> Editar</a>";
+                                    }
+                                    if ($this->data['button']['delete_access_level']) {
+                                        echo "<a href='" . URLADM . "delete-access-level/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")'><i class='fa-regular fa-trash-can'></i> Apagar</a>";
+                                    }
+                                    if ($this->data['button']['list_permission']) {
+                                        echo "<a href='" . URLADM . "list-permission/index/?level=$id'><i class='fa-solid fa-house-lock'></i> Permissão</a>";
+                                    }
+                                    if ($this->data['button']['order_access_level']) {
+                                        echo "<a href='" . URLADM . "order-access-level/index/$id?pag=" . $this->data['pag'] . "'><i class='fa-solid fa-arrow-up-short-wide'></i> Ordem</a>";
+                                    }
                                     ?>
                                 </div>
                             </div>
@@ -65,7 +81,7 @@ if (!defined('G9C8O7N6N5T4I')) {
             </tbody>
         </table>
 
-        <?php $this->data['pagination'] !== [] ? printf($this->data['pagination']) : NULL?>
+        <?php $this->data['pagination'] !== [] ? printf($this->data['pagination']) : NULL ?>
     </div>
 </div>
 <!-- Fim do conteudo do administrativo -->
