+++ app/adms/Views/ConfEmail/ViewConfEmail.php
@@ -11,11 +11,22 @@ if (!defined('G9C8O7N6N5T4I')) {
             <span class="title-content">Detalhes do E-mail</span>
             <div class="top-list-right">
                 <?php
-                echo "<a href='" . URLADM . "list-conf-email/index' class='btn-info'>Listar</a> ";
+                if ($this->data['button']['list_conf_email']) {
+                    echo "<a href='" . URLADM . "list-conf-email/index' class='btn-info'>Listar</a> ";
+                }
+
                 if (!empty($this->data['viewConfEmail'])) {
-                    echo "<a href='" . URLADM . "edit-conf-email/index/" . $this->data['viewConfEmail'][0]['id'] . "' class='btn-warning'>Editar</a> ";
-                    echo "<a href='" . URLADM . "edit-conf-email-password/index/" . $this->data['viewConfEmail'][0]['id'] . "' class='btn-warning'>Editar Senha</a> ";
-                    echo "<a href='" . URLADM . "delete-conf-email/index/" . $this->data['viewConfEmail'][0]['id'] . "' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")' class='btn-danger'>Apagar</a> ";
+                    if ($this->data['button']['edit_conf_email']) {
+                        echo "<a href='" . URLADM . "edit-conf-email/index/" . $this->data['viewConfEmail'][0]['id'] . "' class='btn-warning'>Editar</a> ";
+                    }
+
+                    if ($this->data['button']['edit_conf_email_password']) {
+                        echo "<a href='" . URLADM . "edit-conf-email-password/index/" . $this->data['viewConfEmail'][0]['id'] . "' class='btn-warning'>Editar Senha</a> ";
+                    }
+
+                    if ($this->data['button']['delete_conf_email']) {
+                        echo "<a href='" . URLADM . "delete-conf-email/index/" . $this->data['viewConfEmail'][0]['id'] . "' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")' class='btn-danger'>Apagar</a> ";
+                    }
                 }
                 ?>
             </div>
