+++ app/adms/Views/GroupPages/ListGroupPages.php
@@ -13,7 +13,10 @@ if (!defined('G9C8O7N6N5T4I')) {
             <span class="title-content">Listar Grupos de Páginas</span>
             <div class="top-list-right">
                 <?php
-                echo "<a href='" . URLADM . "add-group-page/index' class='btn-success'>Cadastrar</a>";
+
+                if ($this->data['button']['add_groups_page']) {
+                    echo "<a href='" . URLADM . "add-groups-page/index' class='btn-success'>Cadastrar</a>";
+                }
                 ?>
             </div>
         </div>
@@ -48,10 +51,18 @@ if (!defined('G9C8O7N6N5T4I')) {
                                 <button onclick="actionDropdown(<?php echo $id; ?>)" class="dropdown-btn-action">Ações</button>
                                 <div id="actionDropdown<?php echo $id; ?>" class="dropdown-action-item">
                                     <?php
-                                    echo "<a href='" . URLADM . "order-group-page/index/$id?pag=" . $this->data['pag'] . "'><i class='fa-solid fa-arrow-up-short-wide'></i> Ordem</a>";
-                                    echo "<a href='" . URLADM . "view-group-page/index/$id'><i class='fa-regular fa-eye'></i> Visualizar</a>";
-                                    echo "<a href='" . URLADM . "edit-group-page/index/$id'><i class='fa-regular fa-pen-to-square'></i> Editar</a>";
-                                    echo "<a href='" . URLADM . "delete-group-page/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")'><i class='fa-regular fa-trash-can'></i> Apagar</a>";
+                                    if ($this->data['button']['order_group_page']) {
+                                        echo "<a href='" . URLADM . "order-group-page/index/$id?pag=" . $this->data['pag'] . "'><i class='fa-solid fa-arrow-up-short-wide'></i> Ordem</a>";
+                                    }
+                                    if ($this->data['button']['view_group_page']) {
+                                        echo "<a href='" . URLADM . "view-group-page/index/$id'><i class='fa-regular fa-eye'></i> Visualizar</a>";
+                                    }
+                                    if ($this->data['button']['edit_group_page']) {
+                                        echo "<a href='" . URLADM . "edit-group-page/index/$id'><i class='fa-regular fa-pen-to-square'></i> Editar</a>";
+                                    }
+                                    if ($this->data['button']['delete_group_page']) {
+                                        echo "<a href='" . URLADM . "delete-group-page/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")'><i class='fa-regular fa-trash-can'></i> Apagar</a>";
+                                    }
                                     ?>
                                 </div>
                             </div>
