+++ app/adms/Views/Pages/ListPages.php
@@ -13,8 +13,13 @@ if (!defined('G9C8O7N6N5T4I')) {
             <span class="title-content">Listar Páginas</span>
             <div class="top-list-right">
                 <?php
-                echo "<a href='" . URLADM . "add-page/index' class='btn-success'>Cadastrar</a> ";
-                echo "<a href='" . URLADM . "sync-pages-levels/index' class='btn-warning'>Sincronizar</a>";
+                if ($this->data['button']['add_page']) {
+                    echo "<a href='" . URLADM . "add-page/index' class='btn-success'>Cadastrar</a> ";
+                }
+
+                if ($this->data['button']['sync_pages_levels']) {
+                    echo "<a href='" . URLADM . "sync-pages-levels/index' class='btn-warning'>Sincronizar</a>";
+                }
                 ?>
             </div>
         </div>
@@ -51,9 +56,17 @@ if (!defined('G9C8O7N6N5T4I')) {
                                 <button onclick="actionDropdown(<?php echo $id; ?>)" class="dropdown-btn-action">Ações</button>
                                 <div id="actionDropdown<?php echo $id; ?>" class="dropdown-action-item">
                                     <?php
-                                    echo "<a href='" . URLADM . "view-page/index/$id'>Visualizar</a>";
-                                    echo "<a href='" . URLADM . "edit-page/index/$id'>Editar</a>";
-                                    echo "<a href='" . URLADM . "delete-page/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")'>Apagar</a>";
+                                    if ($this->data['button']['view_page']) {
+                                        echo "<a href='" . URLADM . "view-page/index/$id'>Visualizar</a>";
+                                    }
+
+                                    if ($this->data['button']['edit_page']) {
+                                        echo "<a href='" . URLADM . "edit-page/index/$id'>Editar</a>";
+                                    }
+
+                                    if ($this->data['button']['delete_page']) {
+                                        echo "<a href='" . URLADM . "delete-page/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")'>Apagar</a>";
+                                    }
                                     ?>
                                 </div>
                             </div>
