+++ app/adms/Views/Pages/viewPages.php
@@ -13,10 +13,21 @@ if (!defined('G9C8O7N6N5T4I')) {
             <span class="title-content">Detalhes da Página</span>
             <div class="top-list-right">
                 <?php
-                echo "<a href='" . URLADM . "list-pages/index' class='btn-info'>Listar</a> ";
+                var_dump($this->data['button']);
+                echo "<br>";
+
+                if ($this->data['button']['list_pages']) {
+                    echo "<a href='" . URLADM . "list-pages/index' class='btn-info'>Listar</a> ";
+                }
+
                 if (!empty($this->data['viewPg'])) {
-                    echo "<a href='" . URLADM . "edit-page/index/" . $this->data['viewPg'][0]['id'] . "' class='btn-warning'>Editar</a> ";
-                    echo "<a href='" . URLADM . "delete-page/index/" . $this->data['viewPg'][0]['id'] . "' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")' class='btn-danger'>Apagar</a> ";
+                    if ($this->data['button']['edit_page']) {
+                        echo "<a href='" . URLADM . "edit-page/index/" . $this->data['viewPg'][0]['id'] . "' class='btn-warning'>Editar</a> ";
+                    }
+
+                    if ($this->data['button']['delete_page']) {
+                        echo "<a href='" . URLADM . "delete-page/index/" . $this->data['viewPg'][0]['id'] . "' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")' class='btn-danger'>Apagar</a> ";
+                    }
                 }
                 ?>
             </div>
