+++ app/adms/Views/SituationPages/viewSituationPages.php
@@ -15,10 +15,17 @@ if (!defined('G9C8O7N6N5T4I')) {
             <span class="title-content">Detalhes da Situação da Página</span>
             <div class="top-list-right">
                 <?php
-                echo "<a href='" . URLADM . "list-situation-pages/index' class='btn-info'>Listar</a> ";
+                if ($this->data['button']['list_situation_pages']) {
+                    echo "<a href='" . URLADM . "list-situation-pages/index' class='btn-info'>Listar</a> ";
+                }
                 if (!empty($this->data['viewSitsPg'])) {
-                    echo "<a href='" . URLADM . "edit-situation-page/index/" . $this->data['viewSitsPg'][0]['id'] . "' class='btn-warning'>Editar</a> ";
-                    echo "<a href='" . URLADM . "delete-situation-page/index/" . $this->data['viewSitsPg'][0]['id'] . "' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")' class='btn-danger'>Apagar</a> ";
+                    if ($this->data['button']['edit_situation_page']) {
+                        echo "<a href='" . URLADM . "edit-situation-page/index/" . $this->data['viewSitsPg'][0]['id'] . "' class='btn-warning'>Editar</a> ";
+                    }
+
+                    if ($this->data['button']['delete_situation_page']) {
+                        echo "<a href='" . URLADM . "delete-situation-page/index/" . $this->data['viewSitsPg'][0]['id'] . "' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")' class='btn-danger'>Apagar</a> ";
+                    }
                 }
                 ?>
             </div>
